QA/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
jira==3.5.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
